refactor: rename createTemplate to createContainer
