+++ src/components/addContent/index.js
@@ -1,16 +1,16 @@
 import AddForm from './addForm/index.js';
 import CurrentProductTable from './currentProductTable/index.js';
-import createTemplate from '../utils/createTemplate.js';
+import createContainer from '../utils/createContainer.js';
 import { HEADING_ADD, HEADING_CURRENT_PRODUCT } from './const.js';
 
 const createAddFormContainer = () => {
-  const container = createTemplate('add-form-container', HEADING_ADD);
+  const container = createContainer('add-form-container', HEADING_ADD);
   container.appendChild(new AddForm().getForm());
   return container;
 };
 
 const createCurrentProductTableContainer = (products) => {
-  const container = createTemplate(
+  const container = createContainer(
     'current-product-table-container',
     HEADING_CURRENT_PRODUCT
   );
